+++ git-ref/tests/file/transaction/prepare_and_commit/delete.rs
@@ -159,7 +159,7 @@ fn delete_reflog_only_of_symbolic_no_deref() -> crate::Result {
     assert!(main.log_exists(&store), "log is untouched, too");
     assert_eq!(
         main.target,
-        head.follow_symbolic(&store, None).expect("a symref")?.target,
+        head.follow(&store, None).expect("a symref")?.target,
         "head points to main"
     );
     Ok(())
@@ -193,7 +193,7 @@ fn delete_reflog_only_of_symbolic_with_deref() -> crate::Result {
     assert!(!main.log_exists(&store), "log is removed");
     assert_eq!(
         main.target,
-        head.follow_symbolic(&store, None).expect("a symref")?.target,
+        head.follow(&store, None).expect("a symref")?.target,
         "head points to main"
     );
     Ok(())