+++ git-ref/CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Breaking
 
 * Replace `transaction::Create` with `transaction::PreviousValue` and remove `transaction::Create`
+* Remove `file::Reference` in favor of `Reference`
 
 ### v0.6.1